+++ src/grpc/hi-lo-bit-segmentation/bigint-to-hi-lo.ts
@@ -1,4 +1,4 @@
-import { H40, H96, H128, H160, H256 } from '../../lib';
+import { H40, H96, H128, H160, H256 } from '../../lib/codegen/grpc/types_pb';
 
 type BigIntable = string | number | bigint | boolean;